Move pretty symbols to a helper class
