+++ packages/phoenix-app/src/app/components/ui-menu/collections-info/collections-info-overlay/collections-info-overlay.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { PhoenixObjects } from 'phoenix-event-display';
+import { PrettySymbols } from 'phoenix-event-display';
 import { EventDisplayService } from 'src/app/services/event-display.service';
 
 @Component({
@@ -30,7 +30,7 @@ export class CollectionsInfoOverlayComponent implements OnInit {
 
   changeCollection(selected: any) {
     const value = selected.target.value;
-    this.showingCollection = this.eventDisplay.getCollection(value).map(PhoenixObjects.getPrettyParams);
+    this.showingCollection = this.eventDisplay.getCollection(value).map(PrettySymbols.getPrettyParams);
     this.collectionColumns = Object.keys(this.showingCollection[0]).filter((column) => column !== 'uuid');
   }
