+++ packages/phoenix-event-display/src/loaders/objects/phoenix-objects.ts
@@ -6,74 +6,6 @@ import { RKHelper } from '../../helpers/rk-helper';
  */
 export class PhoenixObjects {
 
-  /** Pretty symbols for object params. */
-  public static readonly prettySymbols: object = {
-    'θ': ['theta'],
-    'ϕ': ['phi'],
-    'pT': ['pt'],
-    '𝛘2': ['chi2'],
-    'η': ['eta'],
-    'q': ['charge'],
-    'NDOF': ['ndof'],
-    'DOF': ['dof'],
-    'Energy': ['energy'],
-    'ET': ['et'],
-    '|p|': ['momentum', 'mom']
-  };
-
-  /**
-   * Get pretty symbol of a parameter.
-   * @param param Parameter of a physics object.
-   */
-  public static getPrettySymbol(param: string) {
-    const prettySymbol = Object.keys(PhoenixObjects.prettySymbols)
-      .find(symbol => PhoenixObjects.prettySymbols[symbol].includes(param));
-
-    return prettySymbol ? prettySymbol : param;
-  }
-
-  /**
-   * Get pretty printed parameters of an object.
-   * @param params Object parameters to be pretty printed.
-   * @returns New pretty printed parameterss.
-   */
-  public static getPrettyParams(params: { [key: string]: any }): object {
-    // Create a copy of the params so we don't overwrite the original object
-    const paramsCopy = Object.assign({}, params);
-    // Go through all the parameters
-    for (const paramKey of Object.keys(paramsCopy)) {
-      // Get the pretty printed symbol
-      const symbol = PhoenixObjects.getPrettySymbol(paramKey);
-      // If we do get a symbol
-      if (symbol !== paramKey) {
-        // Add a parameter with pretty printed symbol
-        paramsCopy[symbol] = paramsCopy[paramKey];
-        delete paramsCopy[paramKey];
-      }
-    }
-
-    // Delete 'pos' since it's too long and not needed
-    delete paramsCopy['pos'];
-
-    // Pretty print the dparams if any
-    if (paramsCopy?.dparams) {
-      const prettyDParams: object = {};
-
-      prettyDParams['θ'] = paramsCopy.dparams[3];
-      prettyDParams['ϕ'] = paramsCopy.dparams[2];
-      prettyDParams['|p|'] = Math.abs(1 / paramsCopy.dparams[4]);
-      prettyDParams['q'] = Math.sign(1 / paramsCopy.dparams[4]);
-      prettyDParams['d0'] = paramsCopy.dparams[0];
-      prettyDParams['z0'] = paramsCopy.dparams[1];
-
-      delete paramsCopy.dparams;
-
-      return { ...paramsCopy, ...prettyDParams };
-    }
-
-    return paramsCopy;
-  }
-
   /**
    * Process the Track from the given parameters (and positions)
    * and get it as a geometry.
